3hakaton2022/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
python-telegram==0.15.0
    # via -r requirements.in
